pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chaincommand"
version = "2.0.0"
description = "Proactive Virtual Supply Chain AI Team"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "structlog>=23.0",
    "rich>=13.0",
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.2", "langgraph-checkpoint-sqlite>=2.0"]
ortools = ["ortools>=9.8"]
causal = ["dowhy>=0.11", "networkx>=3.0"]
chronos = ["chronos-forecasting>=1.0", "torch>=2.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
    "fastapi>=0.110",
    "uvicorn>=0.29",
]
all = [
    "chaincommand[langgraph,ortools,causal,chronos,dev]",
]

[tool.setuptools.packages.find]
include = ["chaincommand*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B"]
ignore = ["S101", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
